# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nutrition Clinic Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctors/{doctorId}/availability:
    get:
      operationId: getDoctorAvailability
      tags: [doctors]
      summary: Get doctor available slots
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, confirmed, completed, cancelled, no_show]
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: doctorId
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentsListResponse"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status or details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentRequest"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: cancelAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    AppointmentType:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        isActive:
          type: boolean
      required:
        - id
        - nameAr
        - nameEn
        - durationMinutes
        - isActive
    Service:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        description:
          type: string
        price:
          type: number
        vatRate:
          type: number
          description: VAT rate (0.15 for Saudi 15% VAT)
        isActive:
          type: boolean
      required:
        - id
        - nameAr
        - nameEn
        - price
        - vatRate
        - isActive
    Doctor:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        specialization:
          type: string
        phone:
          type: string
        email:
          type: string
        bio:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - nameAr
        - nameEn
        - isActive
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        isAvailable:
          type: boolean
      required:
        - startTime
        - endTime
        - isAvailable
    Patient:
      type: object
      properties:
        id:
          type: string
        nameAr:
          type: string
        nameEn:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female]
        nationalId:
          type: string
        medicalHistory:
          type: string
        allergies:
          type: string
        metadata:
          type: object
          description: Flexible extra fields (weight, height, measurements, etc.)
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - nameAr
        - phone
        - isActive
        - createdAt
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        patientId:
          type: string
        doctorId:
          type: string
        serviceId:
          type: string
        appointmentTypeId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        mode:
          type: string
          enum: [in_person, online]
        notes:
          type: string
        videoLink:
          type: string
        patient:
          $ref: "#/components/schemas/Patient"
        doctor:
          $ref: "#/components/schemas/Doctor"
        service:
          $ref: "#/components/schemas/Service"
        appointmentType:
          $ref: "#/components/schemas/AppointmentType"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - scheduledAt
        - status
        - mode
        - createdAt
    CreateAppointmentRequest:
      type: object
      properties:
        patientId:
          type: string
        doctorId:
          type: string
        serviceId:
          type: string
        appointmentTypeId:
          type: string
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          default: 60
        mode:
          type: string
          enum: [in_person, online]
        notes:
          type: string
      required:
        - patientId
        - scheduledAt
        - mode
    UpdateAppointmentRequest:
      type: object
      properties:
        scheduledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        mode:
          type: string
          enum: [in_person, online]
        notes:
          type: string
        videoLink:
          type: string
    AppointmentsListResponse:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - appointments
        - total
        - page
        - limit
...