# lib/api-zod/src/generated/types/appointmentsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nutrition Clinic Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";

export interface AppointmentsListResponse {
  appointments: Appointment[];
  total: number;
  page: number;
  limit: number;
}