# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nutrition Clinic Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: string;
  nameAr: string;
  nameEn: string;
  specialization?: string;
  phone?: string;
  email?: string;
  bio?: string;
  isActive: boolean;
}